Knoze Stuff

+++ src/main/java/com/rumaruka/thaumicbases/common/block/BlockRedlonStem.java
@@ -185,7 +185,7 @@ public class BlockRedlonStem  extends BlockBush implements IGrowable
      */
     public boolean canGrow(World worldIn, BlockPos pos, IBlockState state, boolean isClient)
     {
-        return ((Integer)state.getValue(AGE)).intValue() != 7;
+        return (Integer) state.getValue(AGE) != 7;
     }
 
     public boolean canUseBonemeal(World worldIn, Random rand, BlockPos pos, IBlockState state)
